refactor: Enhance Token related Middleware Functions
This commit delivers a series of refinements aimed at improving code maintainability:

1- Introduced the jwt-auth.ts file to centralize JWT-related operations, eliminating redundancy in retrieving user information from tokens.

2- Incorporated two new middleware functions, one to verify if a user is nominated and another to check if a user is a moderator.

3-  Migrated the contents of is-auth.ts into jwt-auth.ts
