+++ src/middlewares/index.ts
@@ -1,3 +1,3 @@
 export * from './errorHandling';
 export * from './Multer';
-export * from './is-auth';
+export * from './jwt-auth';
